requirements.txt
# pip install -r requirements.txt

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
opencv-python-headless>=4.5.5.64
opencv-contrib-python>=4.5.5.64
albumentations>=1.1.0
